Clean up the JobStatusListener signature
Also break out Job to its own file to avoid import problems.
