+++ src/bygg/core/scheduler.py
@@ -1,35 +1,13 @@
 from pathlib import Path
 from typing import Literal
 
-from bygg.core.action import Action, CommandStatus
+from bygg.core.action import Action
 from bygg.core.cache import Cache
 from bygg.core.dag import Dag, create_dag
 from bygg.core.digest import calculate_dependency_digest, calculate_digest
+from bygg.core.job import Job
 from bygg.logging import logger
 from bygg.output.status_display import on_check_failed
-
-
-class Job:
-    name: str
-    action: Action
-    status: CommandStatus | None
-
-    def __init__(self, action: Action):
-        self.name = action.name
-        self.action = action
-        self.status = None
-
-    def __repr__(self) -> str:
-        return f'"{self.name}, status: {self.status.rc if self.status else "unknown"}"'
-
-    def __hash__(self) -> int:
-        return hash(self.name)
-
-    def __str__(self) -> str:
-        return self.name
-
-    def __eq__(self, __o: object) -> bool:
-        return isinstance(__o, Job) and self.name == __o.name
 
 
 class Scheduler:
